moved fee payment page to pages
